+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix Default group manual usage ignored in strict groups mode
+
 ### 3.0.0 (2021-01-20)
 * (breaking) Switch to `jakarta.validation` (EE9, Bean validation 3, hibernate-validator 7 support) from `javax.validation`) 
 * Fix module name (to ru.vyarus.guice.validator)